Extract log action
